Refactor server implementation and introduce client tool
Reorganized the server code to use the MCP library for tools, prompts, and resources management, simplifying the architecture. Added a new WeatherTool handler and updated the main function for better modularity. Introduced a new test_client to handle event streams and JSON-RPC communication for initialization and notifications.
